requirements.txt
fastapi==0.95.2
pydantic==1.10.15
uvicorn[standard]==0.24.0
python-dateutil==2.8.2
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi